Move BlockchainService to contracts
